Fix for 64-bit platforms. Probably better do the other way round (aka guint -> gsize) since underlying g_io_channel_read()/... operate the same way. This would change cddb-slave2 visible API. * cddb-slave2/cddb-slave.c: (do_goodbye), (do_read), (do_query), (do_hello): * cddb-slave2/iochannel.c: (gnet_io_channel_writen), (gnet_io_channel_readn): Use guint instead of gsize to comply with cddb-slave API.

+++ cddb-slave2/cddb-slave-capplet.c
@@ -245,7 +245,7 @@ other_port_changed (GtkEntry *entry,
 static void
 do_goodbye (PropertyDialog *pd)
 {
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 	
 	status = gnet_io_channel_writen (pd->iochannel, "quit\n",
@@ -360,7 +360,7 @@ do_sites_response (PropertyDialog *pd,
 static void
 do_sites (PropertyDialog *pd)
 {
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 
 	status = gnet_io_channel_writen (pd->iochannel, "sites\n",
@@ -410,7 +410,7 @@ static void
 do_hello (PropertyDialog *pd)
 {
 	char *hello;
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 
 	/* Use the gconf values */
